# contracts/openapi/argus-api.yaml
openapi: 3.1.0
info:
  title: ARGUS Platform API
  version: 0.1.0
  description: GENERATED by scripts/contracts.py export from backend/app. Do not hand-edit.
paths:
...
  /api/v1/tasks:
    post:
      tags:
      - tasks
      summary: Create Task
      operationId: create_task_api_v1_tasks_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/tasks/{task_id}:
    get:
      tags:
      - tasks
      summary: Get Task
      operationId: get_task_api_v1_tasks__task_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - tasks
      summary: Update Task
      operationId: update_task_api_v1_tasks__task_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sites/{site_id}/tasks:
    get:
      tags:
      - tasks
      summary: List Site Tasks
      operationId: list_site_tasks_api_v1_sites__site_id__tasks_get
      security:
      - HTTPBearer: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          title: Site Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - PLANNED
            - ASSIGNED
            - IN_PROGRESS
            - PAUSED
            - COMPLETED
            - CANCELLED
            type: string
          - type: 'null'
          title: Status
      - name: date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskDTO'
                title: Response List Site Tasks Api V1 Sites  Site Id  Tasks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}/assign:
    post:
      tags:
      - assignments
      summary: Assign Task
      operationId: assign_task_api_v1_tasks__task_id__assign_post
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/assignments:
    post:
      tags:
      - assignments
      summary: Create Assignment
      operationId: create_assignment_api_v1_assignments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/assignments/{assignment_id}:
    patch:
      tags:
      - assignments
      summary: Patch Assignment
      operationId: patch_assignment_api_v1_assignments__assignment_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: assignment_id
        in: path
        required: true
        schema:
          type: string
          title: Assignment Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/operators/{operator_id}/tasks/today:
    get:
      tags:
      - operator-day
      summary: Operator Tasks Today
      operationId: operator_tasks_today_api_v1_operators__operator_id__tasks_today_get
      security:
      - HTTPBearer: []
      parameters:
      - name: operator_id
        in: path
        required: true
        schema:
          type: string
          title: Operator Id
      - name: date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodayTask'
                title: Response Operator Tasks Today Api V1 Operators  Operator Id  Tasks Today Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/operators/{operator_id}/machine/confirm:
    post:
      tags:
      - operator-day
      summary: Confirm Machine
      operationId: confirm_machine_api_v1_operators__operator_id__machine_confirm_post
      security:
      - HTTPBearer: []
      parameters:
      - name: operator_id
        in: path
        required: true
        schema:
          type: string
          title: Operator Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmMachineRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmMachineResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/machines/{machine_id}/precheck:
    get:
      tags:
      - precheck
      summary: Get Precheck
      operationId: get_precheck_api_v1_machines__machine_id__precheck_get
      security:
      - HTTPBearer: []
      parameters:
      - name: machine_id
        in: path
        required: true
        schema:
          type: string
          title: Machine Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecheckTemplateDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - precheck
      summary: Post Precheck
      operationId: post_precheck_api_v1_machines__machine_id__precheck_post
      security:
      - HTTPBearer: []
      parameters:
      - name: machine_id
        in: path
        required: true
        schema:
          type: string
          title: Machine Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrecheckSubmit'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrecheckResultDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/tasks/{task_id}/summary:
    get:
      tags:
      - task-lifecycle
      summary: Task Summary
      operationId: task_summary_api_v1_tasks__task_id__summary_get
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/telemetry:
    post:
      tags:
      - telemetry
      summary: Post Telemetry
      operationId: post_telemetry_api_v1_telemetry_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryBatch'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/machines/{machine_id}/telemetry:
    get:
      tags:
      - telemetry
      summary: Get Machine Telemetry
      operationId: get_machine_telemetry_api_v1_machines__machine_id__telemetry_get
      security:
      - HTTPBearer: []
      parameters:
      - name: machine_id
        in: path
        required: true
        schema:
          type: string
          title: Machine Id
      - name: from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: From
      - name: to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: To
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 5000
          default: 500
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetryPoint'
                title: Response Get Machine Telemetry Api V1 Machines  Machine Id  Telemetry Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AssignRequest:
      properties:
        operator_id:
          type: string
          title: Operator Id
        machine_id:
          type: string
          title: Machine Id
        replace:
          type: boolean
          title: Replace
          description: Cancel the current active assignment and reassign
          default: false
      type: object
      required:
      - operator_id
      - machine_id
      title: AssignRequest
    AssignmentCreate:
      properties:
        operator_id:
          type: string
          title: Operator Id
        machine_id:
          type: string
          title: Machine Id
        replace:
          type: boolean
          title: Replace
          description: Cancel the current active assignment and reassign
          default: false
        task_id:
          type: string
          title: Task Id
      type: object
      required:
      - operator_id
      - machine_id
      - task_id
      title: AssignmentCreate
    AssignmentDTO:
      properties:
        assignment_id:
          type: string
          title: Assignment Id
        task_id:
          type: string
          title: Task Id
        operator_id:
          type: string
          title: Operator Id
        machine_id:
          type: string
          title: Machine Id
        site_id:
          type: string
          title: Site Id
        assigned_by:
          type: string
          title: Assigned By
        status:
          type: string
          enum:
          - ACTIVE
          - CANCELLED
          - COMPLETED
          title: Status
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - assignment_id
      - task_id
      - operator_id
      - machine_id
      - site_id
      - assigned_by
      - status
      title: AssignmentDTO
    AssignmentUpdate:
      properties:
        status:
          anyOf:
          - type: string
            pattern: ^CANCELLED$
          - type: 'null'
          title: Status
        machine_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Machine Id
      type: object
      title: AssignmentUpdate
    AssistanceItem:
      properties:
        kind:
          type: string
          enum:
          - MACHINE
          - PERSON
          - RESOURCE
          - NOTE
          title: Kind
        resource_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Id
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - kind
      title: AssistanceItem
...
    ConfirmMachineRequest:
      properties:
        machine_id:
          type: string
          title: Machine Id
      type: object
      required:
      - machine_id
      title: ConfirmMachineRequest
    ConfirmMachineResponse:
      properties:
        operator_id:
          type: string
          title: Operator Id
        machine_id:
          type: string
          title: Machine Id
        confirmed_at:
          type: string
          format: date-time
          title: Confirmed At
      type: object
      required:
      - operator_id
      - machine_id
      - confirmed_at
      title: ConfirmMachineResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IngestResult:
      properties:
        received:
          type: integer
          title: Received
        accepted:
          type: integer
          title: Accepted
        duplicates:
          type: integer
          title: Duplicates
      type: object
      required:
      - received
      - accepted
      - duplicates
      title: IngestResult
...
    PrecheckResultDTO:
      properties:
        precheck_id:
          type: string
          title: Precheck Id
        machine_id:
          type: string
          title: Machine Id
        operator_id:
          type: string
          title: Operator Id
        template_id:
          type: string
          title: Template Id
        passed:
          type: boolean
          title: Passed
        failed_items:
          items:
            type: string
          type: array
          title: Failed Items
        results:
          items:
            $ref: '#/components/schemas/PrecheckAnswer'
          type: array
          title: Results
        submitted_at:
          type: string
          format: date-time
          title: Submitted At
      type: object
      required:
      - precheck_id
      - machine_id
      - operator_id
      - template_id
      - passed
      - failed_items
      - results
      - submitted_at
      title: PrecheckResultDTO
    PrecheckSubmit:
      properties:
        results:
          items:
            $ref: '#/components/schemas/PrecheckAnswer'
          type: array
          title: Results
      type: object
      required:
      - results
      title: PrecheckSubmit
    PrecheckTemplateDTO:
      properties:
        template_id:
          type: string
          title: Template Id
        machine_type:
          type: string
          title: Machine Type
        version:
          type: integer
          title: Version
        items:
          items:
            $ref: '#/components/schemas/PrecheckItem'
          type: array
          title: Items
      type: object
      required:
      - template_id
      - machine_type
      - version
      - items
      title: PrecheckTemplateDTO
...
    TaskCreate:
      properties:
        task_id:
          anyOf:
          - type: string
            pattern: ^[A-Z0-9_]{2,32}$
          - type: 'null'
          title: Task Id
          description: Optional; generated if omitted
        site_id:
          type: string
          title: Site Id
        zone_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone Id
        title:
          type: string
          maxLength: 160
          minLength: 1
          title: Title
        task_type:
          type: string
          enum:
          - EXCAVATION
          - TRENCHING
          - LOADING
          - HAULING
          - GRADING
          - BACKFILLING
          - OTHER
          title: Task Type
        priority:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - CRITICAL
          title: Priority
          default: MEDIUM
        target:
          $ref: '#/components/schemas/TaskTarget'
        planned_start:
          type: string
          format: date-time
          title: Planned Start
        planned_end:
          type: string
          format: date-time
          title: Planned End
        assistance:
          items:
            $ref: '#/components/schemas/AssistanceItem'
          type: array
          title: Assistance
          default: []
      type: object
      required:
      - site_id
      - title
      - task_type
      - target
      - planned_start
      - planned_end
      title: TaskCreate
    TaskDTO:
      properties:
        task_id:
          type: string
          title: Task Id
        site_id:
          type: string
          title: Site Id
        zone_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone Id
        title:
          type: string
          title: Title
        task_type:
          type: string
          enum:
          - EXCAVATION
          - TRENCHING
          - LOADING
          - HAULING
          - GRADING
          - BACKFILLING
          - OTHER
          title: Task Type
        priority:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - CRITICAL
          title: Priority
        target:
          $ref: '#/components/schemas/TaskTarget'
        planned_start:
          type: string
          format: date-time
          title: Planned Start
        planned_end:
          type: string
          format: date-time
          title: Planned End
        assistance:
          items:
            $ref: '#/components/schemas/AssistanceItem'
          type: array
          title: Assistance
          default: []
        status:
          type: string
          enum:
          - PLANNED
          - ASSIGNED
          - IN_PROGRESS
          - PAUSED
          - COMPLETED
          - CANCELLED
          title: Status
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - task_id
      - site_id
      - title
      - task_type
      - priority
      - target
      - planned_start
      - planned_end
      - status
      title: TaskDTO
    TaskSummary:
      properties:
        task:
          $ref: '#/components/schemas/TaskDTO'
        session:
          anyOf:
          - $ref: '#/components/schemas/SessionDTO'
          - type: 'null'
        planned_duration_min:
          type: number
          title: Planned Duration Min
        active_duration_min:
          anyOf:
          - type: number
          - type: 'null'
          title: Active Duration Min
        pause_duration_min:
          anyOf:
          - type: number
          - type: 'null'
          title: Pause Duration Min
        cycles_done:
          anyOf:
          - type: integer
          - type: 'null'
          title: Cycles Done
        fuel_used_pct:
          anyOf:
          - type: number
          - type: 'null'
          title: Fuel Used Pct
        idle_min:
          anyOf:
          - type: number
          - type: 'null'
          title: Idle Min
        finished_by_deadline:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Finished By Deadline
      type: object
      required:
      - task
      - planned_duration_min
      title: TaskSummary
    TaskTarget:
      properties:
        unit:
          type: string
          enum:
          - cycles
          - m3
          - tonnes
          - minutes
          title: Unit
        value:
          type: number
          exclusiveMinimum: 0.0
          title: Value
      type: object
      required:
      - unit
      - value
      title: TaskTarget
    TaskUpdate:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 160
            minLength: 1
          - type: 'null'
          title: Title
        zone_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone Id
        priority:
          anyOf:
          - type: string
            enum:
            - LOW
            - MEDIUM
            - HIGH
            - CRITICAL
          - type: 'null'
          title: Priority
        target:
          anyOf:
          - $ref: '#/components/schemas/TaskTarget'
          - type: 'null'
        planned_start:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Planned Start
        planned_end:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Planned End
          description: The task deadline
        assistance:
          anyOf:
          - items:
              $ref: '#/components/schemas/AssistanceItem'
            type: array
          - type: 'null'
          title: Assistance
        status:
          anyOf:
          - type: string
            enum:
            - PLANNED
            - ASSIGNED
            - IN_PROGRESS
            - PAUSED
            - COMPLETED
            - CANCELLED
          - type: 'null'
          title: Status
          description: Only CANCELLED is accepted here; lifecycle uses the action endpoints
      type: object
      title: TaskUpdate
    TelemetryBatch:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TelemetryPoint'
          type: array
          maxItems: 5000
          minItems: 1
          title: Items
      type: object
      required:
      - items
      title: TelemetryBatch
    TelemetryPoint:
      properties:
        machine_id:
          type: string
          title: Machine Id
        ts:
          type: string
          format: date-time
          title: Ts
        engine_hours:
          type: number
          minimum: 0.0
          title: Engine Hours
        fuel_pct:
          type: number
          maximum: 100.0
          minimum: 0.0
          title: Fuel Pct
        fuel_rate_lph:
          type: number
          minimum: 0.0
          title: Fuel Rate Lph
        load_cycles:
          type: integer
          minimum: 0.0
          title: Load Cycles
          description: Cumulative load-cycle counter
        idle_seconds:
          type: number
          minimum: 0.0
          title: Idle Seconds
          description: Cumulative idle seconds counter
        speed_kmh:
          type: number
          minimum: 0.0
          title: Speed Kmh
        rpm:
          type: number
          minimum: 0.0
          title: Rpm
        temperature_c:
          type: number
          title: Temperature C
        lat:
          type: number
          maximum: 90.0
          minimum: -90.0
          title: Lat
        lon:
          type: number
          maximum: 180.0
          minimum: -180.0
          title: Lon
        operating_mode:
          type: string
          enum:
          - DIGGING
          - SWINGING
          - DUMPING
          - LOADING
          - HAULING
          - RETURNING
          - TRAVELLING
          - IDLE
          - 'OFF'
          title: Operating Mode
        seatbelt:
          type: boolean
          title: Seatbelt
      type: object
      required:
      - machine_id
      - ts
      - engine_hours
      - fuel_pct
      - fuel_rate_lph
      - load_cycles
      - idle_seconds
      - speed_kmh
      - rpm
      - temperature_c
      - lat
      - lon
      - operating_mode
      - seatbelt
      title: TelemetryPoint
    TodayTask:
      properties:
        task:
          $ref: '#/components/schemas/TaskDTO'
        assignment:
          $ref: '#/components/schemas/AssignmentDTO'
        session:
          anyOf:
          - $ref: '#/components/schemas/SessionDTO'
          - type: 'null'
        allowed_actions:
          items:
            type: string
          type: array
          title: Allowed Actions
      type: object
      required:
      - task
      - assignment
      - allowed_actions
      title: TodayTask
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer